// src/app/dashboard/op-maintenance/op-maintenance-details/op-auth-plan/model/rpdma67.model.ts
export class Rpdma67 {
  m67tit1 = '';
  m67tit2 = '';
  m67lit1 = '';
  m67val1 = '';
  m67opos = '';
  m67pdes = '';
  m67mhdr = '';
  m67amed = '';
  m67mi1 = '';
  m67ms1 = '';
  m67mi2 = '';
  m67ms2 = '';
  m67mi3 = '';
  m67ms3 = '';
  m67mi4 = '';
  m67ms4 = '';
  m67mi5 = '';
  m67ms5 = '';
  m67mi6 = '';
  m67ms6 = '';
  m67mi7 = '';
  m67ms7 = '';
  m67mi8 = '';
  m67ms8 = '';
  m67mi9 = '';
  m67ms9 = '';
  m67mi10 = '';
  m67ms10 = '';
  m67mi11 = '';
  m67ms11 = '';
  m67mi12 = '';
  m67ms12 = '';
  m67mi13 = '';
  m67ms13 = '';
  m67mi14 = '';
  m67ms14 = '';
  m67mi15 = '';
  m67ms15 = '';
  m67mi16 = '';
  m67ms16 = '';
  m67mi17 = '';
  m67ms17 = '';
  m67mi18 = '';
  m67ms18 = '';
  m67mi19 = '';
  m67ms19 = '';
  m67mi20 = '';
  m67ms20 = '';
  m67mi21 = '';
  m67ms21 = '';
  m67mi22 = '';
  m67ms22 = '';
  m67mi23 = '';
  m67ms23 = '';
  m67mi24 = '';
  m67ms24 = '';
  m67mi25 = '';
  m67ms25 = '';
  m67mi26 = '';
  m67ms26 = '';
  m67mi27 = '';
  m67ms27 = '';
  m67mi28 = '';
  m67ms28 = '';
  m67mi29 = '';
  m67ms29 = '';
  m67mi30 = '';
  m67ms30 = '';
  m67mi31 = '';
  m67ms31 = '';
  m67mi32 = '';
  m67ms32 = '';
  m67mi33 = '';
  m67ms33 = '';
  m67mi34 = '';
  m67ms34 = '';
  m67mi35 = '';
  m67ms35 = '';
  m67mi36 = '';
  m67ms36 = '';
  m67mi37 = '';
  m67ms37 = '';
  m67mi38 = '';
  m67ms38 = '';
  m67mi39 = '';
  m67ms39 = '';
  m67mi40 = '';
  m67ms40 = '';
  m67mi41 = '';
  m67ms41 = '';
  m67mi42 = '';
  m67ms42 = '';
  m67mi43 = '';
  m67ms43 = '';
  m67mi44 = '';
  m67ms44 = '';
  m67mi45 = '';
  m67ms45 = '';
  m67mi46 = '';
  m67ms46 = '';
  m67mi47 = '';
  m67ms47 = '';
  m67mi48 = '';
  m67ms48 = '';
  m67mi49 = '';
  m67ms49 = '';
  m67mi50 = '';
  m67ms50 = '';
  m67mi51 = '';
  m67ms51 = '';
  m67mi52 = '';
  m67ms52 = '';
  m67mi53 = '';
  m67ms53 = '';
  m67mi54 = '';
  m67ms54 = '';
  m67mi55 = '';
  m67ms55 = '';
  m67mi56 = '';
  m67ms56 = '';
  m67mi57 = '';
  m67ms57 = '';
  m67mi58 = '';
  m67ms58 = '';
  m67mi59 = '';
  m67ms59 = '';
  m67mi60 = '';
  m67ms60 = '';
  m67mi61 = '';
  m67ms61 = '';
  m67mi62 = '';
  m67ms62 = '';
  m67mi63 = '';
  m67ms63 = '';
  m67mi64 = '';
  m67ms64 = '';
  m67mi65 = '';
  m67ms65 = '';
  m67hhdr = '';
  m67ahip = '';
  m67hi1 = '';
  m67hp1 = '';
  m67hi2 = '';
  m67hp2 = '';
  m67hi3 = '';
  m67hp3 = '';
  m67hi4 = '';
  m67hp4 = '';
  m67hi5 = '';
  m67hp5 = '';
  m67hi6 = '';
  m67hp6 = '';
  m67hi7 = '';
  m67hp7 = '';
  m67hi8 = '';
  m67hp8 = '';
  m67hi9 = '';
  m67hp9 = '';
  m67hi10 = '';
  m67hp10 = '';
  m67hi11 = '';
  m67hp11 = '';
  m67hi12 = '';
  m67hp12 = '';
  m67hi13 = '';
  m67hp13 = '';
  m67hi14 = '';
  m67hp14 = '';
  m67hi15 = '';
  m67hp15 = '';
  m67hi16 = '';
  m67hp16 = '';
  m67hi17 = '';
  m67hp17 = '';
  m67hi18 = '';
  m67hp18 = '';
  m67hi19 = '';
  m67hp19 = '';
  m67hi20 = '';
  m67hp20 = '';
  m67hi21 = '';
  m67hp21 = '';
  m67hi22 = '';
  m67hp22 = '';
  m67hi23 = '';
  m67hp23 = '';
  m67hi24 = '';
  m67hp24 = '';
  m67hi25 = '';
  m67hp25 = '';
  m67hi26 = '';
  m67hp26 = '';
  m67hi27 = '';
  m67hp27 = '';
  m67hi28 = '';
  m67hp28 = '';
  m67hi29 = '';
  m67hp29 = '';
  m67hi30 = '';
  m67hp30 = '';
  m67hi31 = '';
  m67hp31 = '';
  m67hi32 = '';
  m67hp32 = '';
  m67hi33 = '';
  m67hp33 = '';
  m67hi34 = '';
  m67hp34 = '';
  m67hi35 = '';
  m67hp35 = '';
  m67hi36 = '';
  m67hp36 = '';
  m67hi37 = '';
  m67hp37 = '';
  m67hi38 = '';
  m67hp38 = '';
  m67hi39 = '';
  m67hp39 = '';
  m67hi40 = '';
  m67hp40 = '';
  m67hi41 = '';
  m67hp41 = '';
  m67hi42 = '';
  m67hp42 = '';
  m67hi43 = '';
  m67hp43 = '';
  m67hi44 = '';
  m67hp44 = '';
  m67hi45 = '';
  m67hp45 = '';
  m67hi46 = '';
  m67hp46 = '';
  m67hi47 = '';
  m67hp47 = '';
  m67hi48 = '';
  m67hp48 = '';
  m67hi49 = '';
  m67hp49 = '';
  m67hi50 = '';
  m67hp50 = '';
  m67hi51 = '';
  m67hp51 = '';
  m67hi52 = '';
  m67hp52 = '';
  m67hi53 = '';
  m67hp53 = '';
  m67hi54 = '';
  m67hp54 = '';
  m67hi55 = '';
  m67hp55 = '';
  m67hi56 = '';
  m67hp56 = '';
  m67hi57 = '';
  m67hp57 = '';
  m67hi58 = '';
  m67hp58 = '';
  m67hi59 = '';
  m67hp59 = '';
  m67hi60 = '';
  m67hp60 = '';
  m67hi61 = '';
  m67hp61 = '';
  m67hi62 = '';
  m67hp62 = '';
  m67hi63 = '';
  m67hp63 = '';
  m67hi64 = '';
  m67hp64 = '';
  m67hi65 = '';
  m67hp65 = '';
  m67err = '';
  m67more = '';
  m67pf4 = '';
  m67pf5 = '';
  map67Title1 = '';
  map67Title2 = '';
  map67Literal = '';
  map67Value = '';
  map67Position = '';
  map67PosDesc = '';
  map67MsHdr = '';
  map67MsAll = '';
  map67MsInfos: Map67MsInfo[] = [];
  map67HipHdr = '';
  map67HipAll = '';
  map67HipInfos: Map67HipInfo[] = [];
  map67Err1 = '';
  map67MoreMsg = '';
  map67Pf4Lit = '';
  map67Pf5Lit = '';
  currentPage = '1';
  msList: [] = [];
  hipList: [] = [];
}